% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuality.R
\docType{methods}
\name{calcCancerCount}
\alias{calcCancerCount}
\alias{calcCancerCount,GeneAberExpr-method}
\title{description of function calcCancerCount}
\usage{
calcCancerCount(res)
}
\arguments{
\item{res}{the result matrix obtained from a IdentifyStates() run.}
}
\description{
Calculate difference between very likely (>0.99) and very unlikely
 (<0.01) p value of the IdentifyStates result matrix.
}
