% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuality.R
\docType{methods}
\name{calcQuality}
\alias{calcQuality}
\alias{calcQuality,GeneAberExpr-method}
\title{description of function calcQuality}
\usage{
calcQuality(model)
}
\arguments{
\item{model}{the model matrix obtained from a GetTestModel() run.}
}
\description{
Calculate the sqared difference between two p value models for the HMM analysis.
}
