% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuality.R
\docType{package}
\name{GeneAberExpr}
\alias{GeneAberExpr}
\title{GeneAberExpr: A package to identify chromosomal aberationd based on single cell expression data.}
\description{
In short the package used the spase matrix most single cell data is stored in.
The main parts are implemented in c++ for speed.
}
\details{
Short way to run the analysis:

myData # is a dgCMatrix containing ordered single cell expression from one chromosome
cancer # the id's of a sample of interesting cells
normal # the ids of a comare group of cells
min=min(myData@x)
if( min > 0 ){ min = 0 } ## no '0's in this object...
max = quantile(myData@x, .99)
model = GetTestModel( myData, seq( min, max,(max - min)/9 ), cancer, normal )
quality = unlist( lapply( seq( 1,nrow(model),2), calcQuality ) )
res = IdentifyStates( chrSums[ sort(order(quality,decreasing=T)[1:800]),] , seq( min, max,(max - min)/9 ) , cancer, normal )
cancerCount = apply( res,2, calcCancerCount )
}
