% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetTestModel}
\alias{GetTestModel}
\alias{GetTestModel,GeneAberExpr-method}
\title{GetTestModel runs sthe estimation only and returns a matrix}
\usage{
GetTestModel(data, range, interest, background, phony)
}
\arguments{
\item{range}{the starting points in the size distribution of the data (10!!)}

\item{interest}{row IDs for the group of interest}

\item{background}{row IDS for the background}

\item{phony}{print status messages (default false)}

\item{X}{the sparse matrix (tests are applied to columns!)}
}
\value{
a matrix with the HMM internals (at the moment)
}
\description{
Mased on the example data a model for each genomic area is developed
Based on these per area models the probability to be in either of these states is calculated.
}
