% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IceCreamTest}
\alias{IceCreamTest}
\alias{IceCreamTest,GeneAberExpr-method}
\title{IceCreamTest processes the Ice cream HMM example}
\usage{
IceCreamTest()
}
\value{
a matrix with the HMM internals (at the moment)
}
\description{
the publication 'An Interactive Spreadsheet for Teaching the Forward-Backward Algorithm'
is a beatuful example of how to calculate an HMM. This function re-runs the excel data.
}
