% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IdentifyStatesTest}
\alias{IdentifyStatesTest}
\alias{IdentifyStatesTest,GeneAberExpr-method}
\title{IdentifyStatesTest Test version of 'IdentifyStates' returing both state probabilites for the first cell.}
\usage{
IdentifyStatesTest(data, range, interest, background)
}
\arguments{
\item{range}{the starting points in the size distribution of the data (10!!)}

\item{interest}{row IDs for the group of interest}

\item{background}{row IDS for the background}

\item{X}{the sparse matrix (tests are applied to columns!)}
}
\value{
a matrix with the probabilites for each region to be in 'interest' state.
}
\description{
Based on the example data a model for each genomic area is developed.
Based on these per area models the probability to be in either of these states is calculated.
}
